requirements.txt
openai
python-dotenv
requests
pytest
streamlit